requirements.txt
# GaiaChat Dependencies
# Install with: pip install -r requirements.txt

# Core LLM
openai>=1.0.0

# Web UI
streamlit>=1.28.0

# Astronomy
astroquery>=0.4.6
astropy>=5.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Configuration
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
